+++ src/vivarium_census_prl_synth_pop/results_processing/ssn_and_itin.py
@@ -2,6 +2,7 @@ from typing import Any, Dict, List
 
 import numpy as np
 import pandas as pd
+from loguru import logger
 from vivarium import Artifact
 from vivarium.framework.randomness import RandomnessStream
 
@@ -41,6 +42,7 @@ def get_simulant_id_maps(
     A dictionary of pd.Series suitable for pd.Series.map, indexed by `address_id`
 
     """
+    logger.info(f"Generating {column_name} maps")
     if column_name != "simulant_id":
         raise ValueError(f"Expected `simulant_id`, got `{column_name}`")
     maps = dict()